Start refactoring eval functions
